Move the test store so that Jest doesn't think it's a test file
