+++ lib/pltr/v2/reducers/__tests__/characters.test.js
@@ -1,4 +1,4 @@
-import { configureStore } from './testStore'
+import { configureStore } from './fixtures/testStore'
 import { emptyFile } from '../../store/newFileState'
 import { loadFile, selectCharacterAttributeBookTab } from '../../actions/ui'
 import {
