+++ lib/pltr/v2/reducers/__tests__/fixtures/testStore.js
@@ -2,9 +2,9 @@ import thunk from 'redux-thunk'
 import { createStore, applyMiddleware } from 'redux'
 import undoable, { excludeAction } from 'redux-undo'
 
-import rootReducer from '../root'
-import { RESET_ACTION_RECORDER, RECORD_LAST_ACTION } from '../../constants/ActionTypes'
-import { SYSTEM_REDUCER_ACTION_TYPES } from '../systemReducers'
+import rootReducer from '../../root'
+import { RESET_ACTION_RECORDER, RECORD_LAST_ACTION } from '../../../constants/ActionTypes'
+import { SYSTEM_REDUCER_ACTION_TYPES } from '../../systemReducers'
 
 export function configureStore(whenClientIsReady, initialState) {
   const reducer = undoable(rootReducer({}), {
